+++ hyperledger/api/__init__.py
@@ -11,6 +11,10 @@
 #    limitations under the License.
 
 
+# flake8: noqa
 from .block import BlockApiMixin
 from .blockchain import BlockChainApiMixin
 from .chaincode import ChainCodeApiMixin
+from .network import NetworkApiMixin
+from .registrar import RegistrarApiMixin
+from .transaction import TransactionApiMixin