+++ hyperledger/client.py
@@ -10,18 +10,18 @@
 #    See the License for the specific language governing permissions and
 #    limitations under the License.
 
+
+import logging
 import requests
 import requests.exceptions
 import six
-import websocket
-
 
 from . import api
 from . import auth
 from . import constants
 from . import errors
 
-from .utils import utils, update_headers, kwargs_from_env
+from .utils import update_headers, kwargs_from_env
 
 
 def from_env(**kwargs):
@@ -32,8 +32,12 @@ class Client(
         requests.Session,
         api.BlockApiMixin,
         api.BlockChainApiMixin,
-        api.ChainCodeApiMixin):
-    def __init__(self, base_url=None, version=constants.DEFAULT_API_VERSION,
+        api.ChainCodeApiMixin,
+        api.NetworkApiMixin,
+        api.RegistrarApiMixin,
+        api.TransactionApiMixin):
+    def __init__(self, base_url=constants.DEFAULT_BASE_URL,
+                 version=constants.DEFAULT_API_VERSION,
                  timeout=constants.DEFAULT_TIMEOUT_SECONDS, tls=False):
         super(Client, self).__init__()
 
@@ -46,6 +50,8 @@ class Client(
         self.timeout = timeout
         self._version = version
         self._auth_configs = auth.load_config()
+
+        self.logger = logging.getLogger(__name__)
 
     def _set_request_timeout(self, kwargs):
         """Prepare the kwargs for an HTTP request by inserting the timeout
@@ -98,7 +104,7 @@ class Client(
 
         args = map(six.moves.urllib.parse.quote_plus, args)
 
-        if kwargs.get('versioned_api', False): # TODO: should be True later
+        if kwargs.get('versioned_api', False):  # TODO: default to True later
             return '{0}/v{1}{2}'.format(
                 self.base_url, self._version, pathfmt.format(*args)
             )