+++ hyperledger/tls.py
@@ -19,7 +19,7 @@ class TLSConfig(object):
         self.assert_fingerprint = assert_fingerprint
 
         # TLS v1.0 seems to be the safest default; SSLv23 fails in mysterious
-        # ways: https://github.com/docker/docker-py/issues/963
+        # ways
 
         self.ssl_version = ssl_version or ssl.PROTOCOL_TLSv1