+++ docs/change_log.md
@@ -1,4 +1,30 @@
 # Change Log
 
+## 0.0.9
+* Link to travis ci for checking.
+* Use tox to do qa.
+
+## 0.0.8
+* Add auth sub module.
+
+## 0.0.7
+* Add more functional tests.
+
+## 0.0.6
+* Add more doc on the API usage.
+
+## 0.0.5
+* Add network, registrar and transaction APIs.
+* Refine the function test case.
+
+## 0.0.4
+* Add the unit test cases.
+
+## 0.0.3
+* Add the chaincode related APIs.
+
+## 0.0.2
+* Add the block and blockchain API.
+
 ## 0.0.1
 * Init the client framework.